% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetDNA2InSilicoTool.R
\name{copyFiles4InsilicoTool}
\alias{copyFiles4InsilicoTool}
\title{copyFiles4InsilicoTool}
\usage{
copyFiles4InsilicoTool(dir_path = ".")
}
\description{
copyFiles4InsilicoTool
}
\author{
Zhiwei Zhou
}
