% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetDNA2InSilicoTool.R
\name{generateFiles4InsilicoMsMs}
\alias{generateFiles4InsilicoMsMs}
\title{generateFiles4InsilicoMsMs}
\usage{
generateFiles4InsilicoMsMs(peak_id, dir_path = ".")
}
\description{
generateFiles4InsilicoMsMs
}
\examples{
generateFiles4InsilicoMsMs
generateFiles4InsilicoMsMs(peak_id = 'M196T420', dir_path = 'G:/00_projects/03_MetDNA2/10_project/MetDNA2_project/Data/20220608_biological_samples/NIST_urine_pos/07_insilico_msms/')
}
\author{
Zhiwei Zhou
}
