% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunTools.R
\name{generateMetFragLocalDB}
\alias{generateMetFragLocalDB}
\title{generateMetFragLocalDB}
\usage{
generateMetFragLocalDB(
  metfrag_candidate,
  base_dir = "./02_metfrag",
  is_output = TRUE
)
}
\arguments{
\item{metfrag_candidate}{the metfrag candidates after random order}
}
\description{
Generate local database for MetFrag
}
\author{
Zhiwei Zhou
}
