% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetDNA2InSilicoTool.R
\name{runCfmIdMatch}
\alias{runCfmIdMatch}
\title{runCfmIdMatch}
\usage{
runCfmIdMatch(
  peak_id,
  dir_path = ".",
  cfmid_path = "F:/software/cfm_id/cfm-id.exe",
  config_file = "F:/software/cfm_id/metab_se_cfm/param_config.txt",
  param_file = "F:/software/cfm_id/metab_se_cfm/param_output0.log",
  score_type = "Jaccard",
  ppm = 25,
  mzabs = 0.01
)
}
\arguments{
\item{cfmid_path}{the path of cfmid program. Default: 'F:/software/cfm_id/cfm-id.exe'}

\item{config_file}{the path of config file}

\item{param_file}{the path of parameter file}

\item{score_type}{score type in CFMID. Default: 'Jaccard'}

\item{ppm}{relative precursor tolerance. Default: 25 ppm}

\item{mzabs}{absolute precursor tolerance. Default: 0.01}
}
\description{
runCfmIdMatch
}
\examples{

}
\author{
Zhiwei Zhou
}
