% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetDNA2InSilicoTool.R
\name{runMetFragMatch}
\alias{runMetFragMatch}
\title{runMetFragMatch}
\usage{
runMetFragMatch(
  peak_id,
  dir_path = ".",
  metfrag_path = "F:/software/metfrag/MetFrag2.4.5-CL.jar",
  ppm = 25,
  mzabs = 0.01,
  frag_ppm = 25
)
}
\arguments{
\item{dir_path}{default: '.'}

\item{ppm}{relative precursor tolerance. Default: 25 ppm}

\item{mzabs}{absolute precursor tolerance. Default: 0.01}

\item{frag_ppm}{relative fragment tolerance. Default: 25 ppm}
}
\description{
this function calls ReSOLUTION package. Please refer ReSOLUTION and MetFrag
}
\examples{

}
\author{
Zhiwei Zhou
}
