% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetDNA2InSilicoTool.R
\name{runMsFinderMatch}
\alias{runMsFinderMatch}
\title{runMsFinderMatch}
\usage{
runMsFinderMatch(
  peak_id,
  dir_path = ".",
  msfinder_path = "F:/software/MSFINDER/MSFINDER_ver_3.24/MsfinderConsoleApp.exe"
)
}
\description{
R interface to run MS-FINDER
}
\author{
Zhiwei Zhou
}
