% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetDNA2InSilicoTool.R
\name{runMsFinderMatch}
\alias{runMsFinderMatch}
\title{runMsFinderMatch}
\usage{
runMsFinderMatch(
  peak_id,
  dir_path = ".",
  msfinder_path = "F:/software/MSFINDER/MSFINDER_ver_3.24/MsfinderConsoleApp.exe"
)
}
\arguments{
\item{peak_id}{peak name}

\item{dir_path}{path of working}

\item{msfinder_path}{msfinder program path}
}
\description{
R interface to run MS-FINDER. If you use this function, please refer MS-FINDER parper.
}
\author{
Zhiwei Zhou
}
